can update mandate or external sale term.
