+++ apps/catalog/catalog/src/app/dashboard/avails/manage/manage.component.scss
@@ -9,12 +9,3 @@
 .arrow_back {
   align-self: flex-start;
 }
-
-bf-form-table{
-  margin: 24px;
-}
-
-section{
-  display: flex;
-  place-content: center;
-}
