+++ apps/catalog/catalog/src/app/dashboard/avails/manage/manage.module.ts
@@ -4,37 +4,19 @@ import { RouterModule } from '@angular/router';
 
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
-import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
-import { MatTooltipModule } from '@angular/material/tooltip';
-import { MatDialogModule } from '@angular/material/dialog';
 
 import { CatalogManageAvailsComponent } from './manage.component';
-import { AvailsFilterModule } from '@blockframes/contract/avails/filter/filter.module';
-import { LanguagesFormModule } from '@blockframes/movie/form/languages/languages.module';
-import { JoinPipeModule, MaxLengthModule, ToGroupLabelPipeModule, ToLabelModule, VersionPipeModule } from '@blockframes/utils/pipes';
-import { FormTableModule } from '@blockframes/ui/form/table/form-table.module';
-
+import { TermFormModule } from '@blockframes/contract/term/components/form/form.module'
 @NgModule({
   declarations: [
     CatalogManageAvailsComponent,
   ],
   imports: [
     CommonModule,
-    ToLabelModule,
-    FormTableModule,
-    MaxLengthModule,
-    JoinPipeModule,
-    AvailsFilterModule,
-    LanguagesFormModule,
-    VersionPipeModule,
-    ToGroupLabelPipeModule,
-
+    TermFormModule,
     //Material
-    MatProgressSpinnerModule,
     MatButtonModule,
     MatIconModule,
-    MatTooltipModule,
-    MatDialogModule,
 
     RouterModule.forChild([{ path: '', component: CatalogManageAvailsComponent }]
     ),
