+++ apps/crm/crm/src/app/crm/pages/contracts/term-form/term-form.component.html
@@ -1,0 +1,5 @@
+<a mat-icon-button routerLink="..">
+  <mat-icon svgIcon="arrow_back"></mat-icon>
+</a>
+<term-form [titleId]="titleId$ | async">
+</term-form>
