+++ apps/crm/crm/src/app/crm/pages/contracts/term-form/term-form.component.scss
@@ -1,0 +1,5 @@
+:host {
+  display: flex;
+  flex-direction: column;
+  margin: 32px;
+}
