+++ libs/contract/src/lib/term/components/form/form.component.scss
@@ -1,0 +1,21 @@
+:host {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: flex-start;
+  margin: 24px;
+}
+
+section{
+  display: flex;
+  place-content: center;
+}
+
+languages-form{
+  margin-bottom: 24px;
+  margin-top: 24px;
+}
+
+button{
+  margin-top: 24px;
+}
